Player Inventory: Rework
== Player Inventory changes ==

A player can now have multiple inventories!

Syntax:
   guns [<inventory position>]

If the player doesn't provide any <inventory position>, the script will
allow the player to edit the current inventory.

The default inventory position is 1 (inventory index = 0).
So, if the player types 'guns' for the first time, they will edit
Inventory 1.

If the player provides an <inventory position>, but is already equipped
with it, the script will allow the player to edit the respective
inventory.

If the player provides an <inventory position> and is not already
equipped with it, the script will equip the respective inventory.

If the player provides an <inventory position> that is lower than 1, the
script will print an error message to the chat window.

If the player provides an <inventory position> that is higher than the
length of their inventory, the script will handle it as the last one,
e.g.:
   if index > len(inventories): index = len(inventories)

== Bug fixes ==

This commit fixes a few bugs regarding certain *_filters for weapon lookups.

== Other fixes ==

With this commit, the way that weapon equipment itself is done is
reworked. If the player chooses a weapon from any weapon menu, the
script will now evaluate if the player already is equipped with it. If
so, then it won't remove and re-equip it, as it was done before.

Signed-off-by: BackRaw <backraw@gmx.de>