config: Add cvar for maximum amount of team changes
The amount of team changes for a player are remembered
and checked against the cvar:

* Player team change count > maximum:
        - client command 'jointeam' blocked entirely

* Player team change count < maximum:
        - increment player team change count
	- respawn after respawn delay
	- allow client command 'jointeam'

Player team change counts are cleared each new round or map change.

This commit also fixes the issue with not being able to join a team in
the middle of a round.

Note: The first time a player joins a team on the server is not remembered.

Signed-off-by: BackRaw <backraw@gmx.de>