Enable auto attaching or detaching the silencer for weapons that can be silenced (CS:GO)
With this commit, the plugin remembers the player's choice for the
silencer. This means, that when a player has detached the silencer using
ATTACK2 (right clicking), the corresponding weapon will be spawned
without the silencer attached.

Each inventory will be handled separately!

Affects the following weapons:
* weapon_m4a1_silencer
* weapon_usp_silencer

It does not work for CS:S, however.

Signed-off-by: BackRaw <backraw@gmx.de>