udm.players.PlayerEntity: Add method refill_ammo()
This commit changes the ammo refill behavior. It now goes as follows:
 Whenever the player presses the reload button, the amount
 of rounds left in the clip is added up the the amount of ammo left.

 Clip     = 5
 Clip Max = 20
 Max Ammo = 100

 On Reload:
  Ammo = Max Ammo + Clip Max - Clip => 100 + 20 - 5 => 115

 The game will do the rest after the reload animation has finished.

The same behavior goes for when the player is shooting an empty clip.
Here, a 'clip fix' value of 1 is added, because we do the check in
weapon_fire (weapon clip still 1).

See https://forums.sourcepython.com/viewtopic.php?f=20&t=1653
for the discussion.

Signed-off-by: BackRaw <backraw@gmx.de>